[WIP] replace vscode built-ins by extension pack
ATM we have one pack, that lists all vscode-builtin
extensions. We could create multiple, giving some
choice about which ones to install or not.

As a first step, let's aim to have more features than
not, and use that pack. For the most part, extensions
listed in a pack should be installed, at the latest
version that is compatible, wr to the vscode extensions
API that the current app supports. So no update of version
needed, and we still get fresh as possible extensions,
every build (assuming eventual build-time resolution of
pack dependencies).

A few of the vscode built-in extensions do not work well,
pass a certain older version. These we need to install the
old way, overriding the pack's newer version. e.g.:
`markdown-language-features` requires v1.39.2

Removed @theia/git in favor of the vscode built-in git. This
permits using popular extensions such as GitLens.

Closes #61
Potentially closes #26

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>